requirements.txt
# Core ML libraries
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.62.0
joblib>=1.1.0
requests>=2.28.0

# Deep Learning
torch>=2.6.0
transformers>=4.40.0
peft>=0.10.0
trl>=0.8.0  # For TRL GRPO training

# Gradient Boosting
lightgbm>=3.3.0

pandas-ta>=0.3.14b0
ta>=0.10.0

# Optional: Hyperparameter optimization
optuna>=3.0.0

# Optional: Advanced visualization
plotly>=5.0.0

# News scraping (optional, for news dataset creation)
playwright>=1.40.0
beautifulsoup4>=4.12.0

# System utilities
psutil>=5.9.0

# Google Cloud Firestore (native GCP client)
google-cloud-firestore>=2.11.0

# Google Cloud Storage
google-cloud-storage>=2.10.0

# Training job orchestration removed - using Vast AI instead
vastai>=0.1.0  # Vast.ai CLI for GPU instance management

# MLflow (for experiment tracking and model registry)
mlflow>=2.0.0

# FastAPI and ML Inference Service
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
prometheus-client>=0.17.0
prometheus-fastapi-instrumentator>=6.1.0
onnxruntime>=1.15.0

# Kafka and Streaming
quixstreams>=0.5.0  # Kafka streaming library
watchdog>=3.0.0  # File system monitoring for job handler

# Database
psycopg2-binary>=2.9.0  # PostgreSQL adapter
